# docs/swagger.yaml
basePath: /api/v1
host: 127.0.0.1:8080
info:
  contact:
    email: me@razeen.me
    name: lupuxiao
    url: https://razeen.me
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://razeen.me
  title: Swagger Example API
  version: "1.0"
paths:
  /file/{id}:
    get:
      consumes:
      - multipart/form-data
      description: 获取文件
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取某个文件
      tags:
      - 文件处理
  /go:
    get:
      consumes:
      - application/json
      description: get record by ID
      parameters:
      - description: userId
        in: path
        name: some_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
  /hello:
    get:
      consumes:
      - multipart/form-data
      description: 向你说Hello
      parameters:
      - description: 人名
        in: query
        name: who
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello Razeen"}'
          schema:
            type: string
        "400":
          description: '{"msg": "who are you"}'
          schema:
            type: string
      summary: 测试SayHello
      tags:
      - 测试
  /json:
    post:
      consumes:
      - application/json
      description: 获取JSON的示例
      produces:
      - application/json
      responses:
        "200":
          description: 返回
      summary: 获取JSON的示例
      tags:
      - JSON
  /list:
    get:
      consumes:
      - multipart/form-data
      description: 文件列表
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 查看文件列表
      tags:
      - 文件处理
  /login:
    post:
      consumes:
      - multipart/form-data
      description: 登入
      parameters:
      - default: admin
        description: 用户名
        in: formData
        name: user
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"login success"}'
          schema:
            type: string
        "400":
          description: '{"msg": "user or password error"}'
          schema:
            type: string
      summary: 登陆
      tags:
      - 登陆
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 上传文件
      tags:
      - 文件处理
swagger: "2.0"